% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstparseR-package.R
\docType{package}
\name{lstparseR-package}
\alias{lstparseR}
\alias{lstparseR-package}
\title{lstparseR: Tidy Parsing of NONMEM Listing Files for Population PK/PD Workflows}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to read, clean, and reshape NONMEM listing (.lst) files into tidy data frames for downstream population pharmacokinetic and pharmacodynamic (PK/PD) analyses. Parsers extract THETA, OMEGA, SIGMA, objective function value (OFV), condition number, and shrinkage values from NONMEM output files. Both runs with and without a covariance step are handled gracefully, returning NA for unavailable quantities rather than raising errors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Clinical-Pharmacy-Saarland-University/lstparseR}
  \item Report bugs at \url{https://github.com/Clinical-Pharmacy-Saarland-University/lstparseR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de}

Authors:
\itemize{
  \item Simeon Rüdesheim \email{simeon.ruedesheim@uni-saarland.de}
  \item Dominik Selzer \email{dominik.selzer@uni-saarland.de}
}

}
\keyword{internal}
