% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lst.R
\name{read_lst_file}
\alias{read_lst_file}
\title{Read a NONMEM Listing File}
\usage{
read_lst_file(path)
}
\arguments{
\item{path}{Character string. Path to the \code{.lst} file.}
}
\value{
A character vector of class \code{"lst"}, where each element is one
line of the file.
}
\description{
Reads a NONMEM \code{.lst} output file into memory as a character vector
with class \code{"lst"} for use by the \verb{fetch_*} family of functions.
}
\examples{
path <- system.file("testdata", "full_cov.lst", package = "lstparsR")
lst  <- read_lst_file(path)
print(lst)
}
